+++ NEWS.md
@@ -15,6 +15,8 @@
   (#922).
 - Fixed a crash when a shared variable is declared inside a instantiated
   generic package in 2019 mode (#923).
+- Fixed a crash when a port with unconstrained array-of-record type is
+  associated with a signal that has a static subtype.
 
 ## Version 1.13.0 - 2024-07-13
 - Type and package generics now work correctly with components and